Adding rest of order mvc tests. Also refactoring domain objects to use lazy fetching + removing notnull annotations from orderItoDTO (order field). It was not needed as order is set during saving in service layer. Bootstrap reduced to single class to make testing easier. Reference #71
