+++ src/main/java/com/blocki/springrestonlinestore/core/domain/OrderItem.java
@@ -15,13 +15,13 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class OrderItem extends  BaseEntity {
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "product_id")
     @ToString.Exclude
     @EqualsAndHashCode.Exclude
     private Product product;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "order_id")
     @ToString.Exclude
     @EqualsAndHashCode.Exclude
