+++ src/main/java/com/blocki/springrestonlinestore/core/domain/Product.java
@@ -17,7 +17,7 @@ public class Product extends BaseEntity {
 
     public enum ProductStatus { AVALIABLE, OUT_OF_STOCK, ERROR }
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "seller_id", nullable = false)
     @ToString.Exclude
     @EqualsAndHashCode.Exclude
